Update unit test - CR deletion and expected result (#57)
As the title says, this PR update test that verifies if POD is deleted with CR deletion. In provided envtest such exact check is not possible becasue garbage collection is not working. In result, test can only verify is OwnerReference is set correctly for POD.

Real test can be done only within real cluster.