ClusterFlowConfig: Remove END action support (#58)
This PR removes check if there is END action in the ClusterFlowConfig custom resource defined. According to specification there will be only one action supported on cluster level - podInterface. Thus, this verification is not needed within controller. Validation if action types are defined correctly within CR is going to be done by the webhook.